requirements.txt
pytorch==1.10.1
torchvision==0.11.2
cudatoolkit==11.3.1
joblib==0.17.0
matplotlib==3.5.3
scipy==1.7.3
tqdm==4.64.1
xgboost==1.5.1
scikit-learn==1.0.2
numpy==1.21.5
pandas==1.3.5
